+++ semcore/data-table/CHANGELOG.md
@@ -2,13 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [16.0.8] - 2025-07-28
+
+### Fixed
+
+- Incorrect type for `selectedRows/onSelectedRowsChange` when `uniqueRowKey` type is not a string.
+
 ## [16.0.7] - 2025-07-23
+
+### Added
+
+- Transition for the table header's `top` attribute.
+- New header prop `animationDuration`.
 
 ### Fixed
 
 - Row hover effect for empty state.
+- Incorrect 'Select all' checkbox behavior with pagination.
 - Header glitches in loading state with sticky header.
-- Incorrect type for `selectedRows/onSelectedRowsChange` when uniqueRowKey is not a string.
 
 ## [16.0.6] - 2025-07-04